refactor: remove imports from original dto init
