+++ src/libecalc/domain/infrastructure/energy_components/legacy_consumer/consumer_function_mapper/pump_consumer_function.py
@@ -1,8 +1,8 @@
 from libecalc.domain.infrastructure.energy_components.legacy_consumer.consumer_function.pump_consumer_function import (
     PumpConsumerFunction,
 )
-from libecalc.domain.process.pump import PumpConsumerFunction as PumpConsumerFunctionDTO
-from libecalc.domain.process.pump import create_pump_model
+from libecalc.domain.process.pump.factory import create_pump_model
+from libecalc.domain.process.pump.pump_consumer_function import PumpConsumerFunction as PumpConsumerFunctionDTO
 
 
 def create_pump_consumer_function(model_dto: PumpConsumerFunctionDTO) -> PumpConsumerFunction:
