+++ src/libecalc/domain/infrastructure/energy_components/legacy_consumer/system/types.py
@@ -1,7 +1,7 @@
 from typing import Union
 
 from libecalc.domain.process.core.compressor.base import CompressorModel
-from libecalc.domain.process.pump import PumpModel
+from libecalc.domain.process.pump.pump import PumpModel
 
 
 class ConsumerSystemComponent:
