+++ src/libecalc/domain/infrastructure/energy_components/pump/component.py
@@ -13,7 +13,7 @@ from libecalc.common.utils.rates import (
 from libecalc.core.result import EcalcModelResult
 from libecalc.core.result import results as core_results
 from libecalc.domain.infrastructure.energy_components.base import BaseConsumerWithoutOperationalSettings
-from libecalc.domain.process.pump import PumpModel
+from libecalc.domain.process.pump.pump import PumpModel
 from libecalc.domain.stream_conditions import Pressure, StreamConditions
