+++ src/libecalc/domain/process/pump/factory.py
@@ -6,7 +6,7 @@ from libecalc.common.chart_type import ChartType
 from libecalc.common.logger import logger
 from libecalc.common.serializable_chart import ChartCurveDTO, SingleSpeedChartDTO, VariableSpeedChartDTO
 from libecalc.domain.process.core.chart import SingleSpeedChart, VariableSpeedChart
-from libecalc.domain.process.pump import PumpModel, PumpModelDTO, PumpSingleSpeed, PumpVariableSpeed
+from libecalc.domain.process.pump.pump import PumpModel, PumpModelDTO, PumpSingleSpeed, PumpVariableSpeed
 
 # def evaluate_streams(
 #     self,
