+++ src/libecalc/dto/__init__.py
@@ -1,41 +1,3 @@
-from libecalc.domain.process.dto import (
-    CompressorChart,
-    CompressorConsumerFunction,
-    CompressorModel,
-    CompressorSampled,
-    CompressorSystemCompressor,
-    CompressorSystemConsumerFunction,
-    CompressorSystemOperationalSetting,
-    CompressorTrainSimplifiedWithKnownStages,
-    CompressorTrainSimplifiedWithUnknownStages,
-    CompressorWithTurbine,
-    ConsumerFunction,
-    DirectConsumerFunction,
-    ElectricEnergyUsageModel,
-    EnergyModel,
-    EnergyModelSampled,
-    FuelEnergyUsageModel,
-    GeneratorSetSampled,
-    GenericChartFromDesignPoint,
-    GenericChartFromInput,
-    PumpSystemConsumerFunction,
-    PumpSystemOperationalSetting,
-    PumpSystemPump,
-    SystemOperationalSetting,
-    TabulatedConsumerFunction,
-    TabulatedData,
-    Turbine,
-    Variables,
-)
-from libecalc.domain.process.dto.compressor import (
-    CompressorStage,
-    InterstagePressureControl,
-    MultipleStreamsCompressorStage,
-    SingleSpeedCompressorTrain,
-    VariableSpeedCompressorTrain,
-    VariableSpeedCompressorTrainMultipleStreamsAndPressures,
-)
-from libecalc.domain.process.pump import PumpConsumerFunction, PumpModelDTO
 from libecalc.dto.emission import Emission
 from libecalc.dto.fuel_type import FuelType
 from libecalc.dto.result_options import ResultOptions
